+++ src/senaite/lims/setuphandlers.py
@@ -44,6 +44,28 @@ def setup_html_filter(portal):
     adapter.style_whitelist = style_whitelist
 
 
+def pre_install(portal_setup):
+    """Runs berfore the first import step of the *default* profile
+
+    This handler is registered as a *pre_handler* in the generic setup profile
+
+    :param portal_setup: SetupTool
+    """
+    logger.info("SENAITE LIMS pre-install handler [BEGIN]")
+
+    # https://docs.plone.org/develop/addons/components/genericsetup.html#custom-installer-code-setuphandlers-py
+    profile_id = "profile-senaite.lims:default"
+    context = portal_setup._getImportContext(profile_id)
+    portal = context.getSite()  # noqa
+
+    # Only install the core once!
+    qi = portal.portal_quickinstaller
+    if not qi.isProductInstalled("bika.lims"):
+        portal_setup.runAllImportStepsFromProfile("profile-bika.lims:default")
+
+    logger.info("SENAITE LIMS pre-install handler [DONE]")
+
+
 def post_install(portal_setup):
     """Runs after the last import step of the *default* profile
 
@@ -51,14 +73,14 @@ def post_install(portal_setup):
 
     :param portal_setup: SetupTool
     """
-    logger.info("SENAITE LIMS install handler [BEGIN]")
+    logger.info("SENAITE LIMS post-install handler [BEGIN]")
 
     # https://docs.plone.org/develop/addons/components/genericsetup.html#custom-installer-code-setuphandlers-py
     profile_id = "profile-senaite.lims:default"
     context = portal_setup._getImportContext(profile_id)
     portal = context.getSite()  # noqa
 
-    logger.info("SENAITE LIMS install handler [DONE]")
+    logger.info("SENAITE LIMS post-install handler [DONE]")
 
 
 def post_uninstall(portal_setup):